:bug: fix: 똑표하기 버튼 클릭시 아코디언 열리고 닫히는 현상 수정
