+++ services/quiz/src/common/models/quizReply.ts
@@ -1,14 +1,16 @@
-export interface QuizReplyByIdResponse {
-  quizReplyHistories: [
-    {
-      quizReplyHistoryId: number;
-      answer: string;
-      likeCount: number;
-      creator: {
-        userId: number;
-        nickname: string;
-        profileImageUrl: string;
-      };
-    }
-  ];
+export interface QuizReplierResPonse {
+  userId: number;
+  nickname: string;
+  profileImageUrl: string;
+}
+
+export interface EachQuizReplyByResponse {
+  quizReplyHistoryId: number;
+  answer: string;
+  likeCount: number;
+  creator: QuizReplierResPonse;
+}
+
+export interface QuizReplyByIdResponses {
+  quizReplyHistories: EachQuizReplyByResponse[];
 }
