+++ services/quiz/src/common/remotes/quizReply.ts
@@ -1,7 +1,7 @@
 import { http } from '@depromeet/http';
 
-import { QuizReplyByIdResponse } from '../models/quizReply';
+import { QuizReplyByIdResponses } from '../models/quizReply';
 
 export async function getQuizReplyById(quizIdParams: string | string[] | undefined) {
-  return await http.get<QuizReplyByIdResponse>(`/api/v1/quiz-reply-histories/quizzes/${quizIdParams}`);
+  return await http.get<QuizReplyByIdResponses>(`/api/v1/quiz-reply-histories/quizzes/${quizIdParams}`);
 }
