+++ services/quiz/src/pages/QuizCheckingPage/remotes/sortingQuizAply.ts
@@ -1,8 +1,8 @@
 import { http } from '@depromeet/http';
-import { QuizReplyByIdResponse } from 'common/models/quizReply';
+import { QuizReplyByIdResponses } from 'common/models/quizReply';
 
 export async function getSortedQuizReplyById(quizIdParams: string | string[] | undefined) {
-  return await http.get<QuizReplyByIdResponse>(
+  return await http.get<QuizReplyByIdResponses>(
     `/api/v1/quiz-reply-histories/quizzes/${quizIdParams}?sort=likeCount&createAt,desc`
   );
 }
