+++ services/quiz/src/pages/QuizVotingPage/components/PeerAnswerItem/style.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const PeerAnswerContainer = styled.div`
+  :first-of-type {
+    margin-top: 0;
+  }
+  margin-top: 12px;
+`;
