--- services/quiz/src/pages/QuizVotingPage/components/PeerAnswerViewer/index.tsx
@@ -1,75 +1,0 @@
-import { theme } from '@depromeet/theme';
-import { Accordion, Text, ToastViewer, UserAvatar } from '@depromeet/toks-components';
-import { Flex, Spacing } from '@toss/emotion-utils';
-import { useRouter } from 'next/router';
-import { useState } from 'react';
-import { useQuery } from 'react-query';
-
-import { DoneNumberNotice } from 'common/components/DoneNumberNotice';
-import { getQuizReplyById } from 'common/remotes/quizReply';
-import { getUser } from 'common/remotes/user';
-import { QUERY_KEYS } from 'constants/queryKeys';
-
-import { PeerAnswerContainer, PeerAnswerWrapper, Wrapper } from './style';
-import { VoteButton } from './VoteButton.tsx';
-
-export function PeerAnswerViewer() {
-  const {
-    query: { quizIdParams },
-  } = useRouter();
-  const [isFold, setIsFold] = useState(true);
-
-  const { data: quizzes } = useQuery(QUERY_KEYS.GET_QUIZREPLIES_BY_ID, () => getQuizReplyById(quizIdParams), {
-    enabled: Boolean(quizIdParams),
-  });
-
-  const { data: user } = useQuery(QUERY_KEYS.GET_USER_INFO, getUser);
-
-  if (!quizzes || !user) {
-    return null;
-  }
-
-  const peerAnswers = quizzes.quizReplyHistories.filter(element => element.creator.nickname !== user.nickname);
-
-  return (
-    <Wrapper>
-      <Spacing size={'5vh'} />
-      <Flex css={{ justifyContent: 'space-between' }}>
-        <Text variant="headline">팀원들의 답안도 확인해볼까요? </Text>
-        <DoneNumberNotice done={peerAnswers.length} />
-      </Flex>
-      <Spacing size={'2vh'} />
-      <PeerAnswerWrapper>
-        {peerAnswers.map(({ quizReplyHistoryId, answer, creator }) => (
-          <PeerAnswerContainer key={creator.userId}>
-            <Accordion
-              isFold={isFold}
-              onFold={() => setIsFold(prev => !prev)}
-              backgroundColor={theme.colors.gray110}
-              accordionStyle={{
-                padding: '22px 16px',
-              }}
-              headerNodes={
-                <Flex css={{ alignItems: 'center', width: '100%', justifyContent: 'space-between' }}>
-                  <Flex css={{ alignItems: 'center' }}>
-                    <UserAvatar image={creator.profileImageUrl} size="large" />
-                    <Text css={{ marginLeft: '12px' }} variant="subhead" color="gray020">
-                      {creator.nickname}
-                    </Text>
-                  </Flex>
-                  <VoteButton quizReplyHistoryId={quizReplyHistoryId} peerAnswers={peerAnswers} />
-                </Flex>
-              }
-              bodyNodes={
-                <>
-                  <Spacing size={22} />
-                  <ToastViewer answer={answer} />
-                </>
-              }
-            />
-          </PeerAnswerContainer>
-        ))}
-      </PeerAnswerWrapper>
-    </Wrapper>
-  );
-}
